requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Opcional - para gerar PDFs
# weasyprint>=60.0
